move periodic actions to asyncaction remove gradle
