+++ app/org/tronscan/actions/AsyncAction.scala
@@ -1,0 +1,7 @@
+package org.tronscan.actions
+
+import scala.concurrent.{ExecutionContext, Future}
+
+trait AsyncAction {
+  def execute(implicit executionContext: ExecutionContext): Future[_]
+}
